Decompose tools_execution.py into three cohesive modules
Split src/autoskillit/server/tools/tools_execution.py (924 lines, 5 MCP tool handlers)
into three modules:

- tools_execution.py (~515 lines): retains run_cmd, run_python, run_skill
- tools_fleet_dispatch.py (~247 lines): dispatch_food_truck, record_gate_dispatch
- tools_execution_helpers.py (~191 lines): _import_and_call, _coerce_scalar

No behavioral changes. All tests pass with updated import paths.
Added structural guard test for decomposition boundaries.
Updated arch test caps (server/tools: 16→18) and layer enforcement scans.
