+++ src/autoskillit/server/tools/tools_execution.py
@@ -3,14 +3,10 @@
 from __future__ import annotations
 
 import asyncio
-import functools
 import json
 import os
 import re
 import time
-import types
-import typing
-from collections.abc import Callable
 from pathlib import Path
 
 import structlog
@@ -30,203 +26,17 @@ from autoskillit.server import mcp
 from autoskillit.server._guards import (
     _check_dry_walkthrough,
     _require_enabled,
-    _require_fleet,
     _require_orchestrator_or_higher,
     _validate_skill_command,
 )
 from autoskillit.server._misc import SCENARIO_STEP_NAME_ENV
 from autoskillit.server._notify import _notify, track_response_size
 from autoskillit.server._subprocess import _run_subprocess
+from autoskillit.server.tools.tools_execution_helpers import _import_and_call
 
 logger = get_logger(__name__)
 
 _RUN_PYTHON_SENTINEL_KEYS: frozenset[str] = frozenset({"callable", "timeout"})
-
-
-def _coerce_scalar(val: object, annotation: object) -> object:
-    """Coerce val to match the annotated type.
-
-    Handles str, int, float, and Optional[T] / T | None variants.
-    Skips containers, unions, bool, and unconvertible values.
-    """
-    if isinstance(val, bool):
-        return val
-
-    actual = annotation
-
-    # Unwrap X | None (types.UnionType — bare union syntax, Python 3.10+)
-    if isinstance(annotation, types.UnionType):
-        non_none = [a for a in annotation.__args__ if a is not type(None)]
-        if len(non_none) == 1:
-            actual = non_none[0]
-    # Unwrap Optional[X] / Union[X, None] (typing.Union with __origin__)
-    elif hasattr(annotation, "__origin__") and hasattr(annotation, "__args__"):
-        ann: typing.Any = annotation  # type: ignore[name-defined]
-        origin = ann.__origin__
-        args = ann.__args__
-        if origin is typing.Union:
-            non_none = [a for a in args if a is not type(None)]
-            if len(non_none) == 1:
-                actual = non_none[0]
-
-    # str ← int/float
-    if actual is str and not isinstance(val, str):
-        if isinstance(val, (int, float)):
-            return str(val)
-        return val
-    # int ← str (try/except for unconvertible)
-    if actual is int and not isinstance(val, int):
-        if isinstance(val, str):
-            try:
-                return int(val)
-            except ValueError:
-                return val
-        return val
-    # float ← str/int (try/except for unconvertible)
-    if actual is float and not isinstance(val, float):
-        if isinstance(val, (str, int)):
-            try:
-                return float(val)
-            except ValueError:
-                return val
-        return val
-    return val
-
-
-async def _import_and_call(
-    dotted_path: str,
-    args: dict[str, object] | None = None,
-    timeout: float = 30,
-) -> dict[str, object]:
-    """Import a Python callable by dotted path and invoke it.
-
-    Returns dict with 'success', 'result' (or 'error').
-    Handles sync and async callables, with timeout protection.
-    """
-    import importlib
-    import inspect
-
-    if args is None:
-        args = {}
-
-    if "." not in dotted_path:
-        return {"success": False, "error": f"Invalid dotted path: {dotted_path!r}"}
-
-    module_path, attr_name = dotted_path.rsplit(".", 1)
-
-    try:
-        module = importlib.import_module(module_path)
-    except ImportError as exc:
-        return {"success": False, "error": f"Import failed for {module_path!r}: {exc}"}
-
-    try:
-        func = getattr(module, attr_name)
-    except AttributeError:
-        return {
-            "success": False,
-            "error": f"Module {module_path!r} has no attribute {attr_name!r}",
-        }
-
-    if not callable(func):
-        return {"success": False, "error": f"{dotted_path!r} is not callable"}
-
-    sig = inspect.signature(func)
-
-    for key in list(args.keys()):
-        if key in _RUN_PYTHON_SENTINEL_KEYS:
-            logger.warning(
-                "run_python stripped sentinel key from args",
-                callable=dotted_path,
-                arg_name=key,
-            )
-            del args[key]
-
-    accepts_var_keyword = any(
-        p.kind == inspect.Parameter.VAR_KEYWORD for p in sig.parameters.values()
-    )
-    if not accepts_var_keyword:
-        valid_keys = set(sig.parameters.keys())
-        for key in list(args.keys()):
-            if key not in valid_keys:
-                logger.warning(
-                    "run_python dropped unrecognized arg",
-                    callable=dotted_path,
-                    arg_name=key,
-                    extra_args=[key],
-                )
-                del args[key]
-
-    try:
-        type_hints = typing.get_type_hints(func)
-    except Exception:
-        logger.warning(
-            "get_type_hints failed, skipping coercion", callable=dotted_path, exc_info=True
-        )
-        type_hints = {}
-    coerced: dict[str, object] = {}
-    for key, val in args.items():
-        if val is None and key in sig.parameters:
-            param = sig.parameters[key]
-            if param.default is not inspect.Parameter.empty and param.default is not None:
-                logger.warning(
-                    "run_python null-arg coerced to default",
-                    callable=dotted_path,
-                    arg=key,
-                    default=repr(param.default),
-                )
-                coerced[key] = param.default
-                continue
-        if val is not None and key in type_hints:
-            annotation = type_hints[key]
-            coerced_val = _coerce_scalar(val, annotation)
-            if coerced_val is not val:
-                logger.warning(
-                    "run_python type coerced",
-                    callable=dotted_path,
-                    arg=key,
-                    from_type=type(val).__name__,
-                    to_type=type(coerced_val).__name__,
-                )
-                coerced[key] = coerced_val
-                continue
-        coerced[key] = val
-    args = coerced
-
-    try:
-        if inspect.iscoroutinefunction(func):
-            result = await asyncio.wait_for(func(**args), timeout=timeout)
-        else:
-            result = await asyncio.wait_for(asyncio.to_thread(func, **args), timeout=timeout)
-    except TimeoutError:
-        logger.warning(
-            "run_python timed out; sync thread may continue running",
-            dotted_path=dotted_path,
-            timeout=timeout,
-        )
-        return {"success": False, "error": f"Timeout after {timeout}s calling {dotted_path}"}
-    except Exception as exc:
-        logger.warning(
-            "run_python execution failed",
-            dotted_path=dotted_path,
-            error=type(exc).__name__,
-        )
-        return {"success": False, "error": f"{type(exc).__name__}: {exc}"}
-
-    try:
-        json.dumps(result)
-        return {"success": True, "result": result}
-    except (TypeError, ValueError):
-        return {"success": True, "result": str(result)}
-
-
-def _get_food_truck_prompt_builder() -> Callable[..., str]:
-    """Return the food truck prompt builder with mcp_prefix pre-bound."""
-    from autoskillit.core import detect_autoskillit_mcp_prefix
-    from autoskillit.fleet import _build_food_truck_prompt
-
-    mcp_prefix = detect_autoskillit_mcp_prefix()
-    return functools.partial(_build_food_truck_prompt, mcp_prefix=mcp_prefix)
-
 
 _PURE_SLEEP_RE = re.compile(
     r'^(?:python3?\s+-c\s+["\']import time;\s*time\.sleep\((?P<py_secs>\d+(?:\.\d+)?)\)["\']'
@@ -703,222 +513,3 @@ async def run_skill(
     except BaseException:
         logger.warning("run_skill cancelled", exc_info=True)
         raise
-
-
-@mcp.tool(
-    tags={"autoskillit", "kitchen-core", "fleet"},
-    annotations={"readOnlyHint": True},
-)
-@track_response_size("dispatch_food_truck")
-async def dispatch_food_truck(
-    recipe: str,
-    task: str,
-    ingredients: dict[str, str] | None = None,
-    dispatch_name: str | None = None,
-    timeout_sec: int | None = None,
-    capture: dict[str, str] | None = None,
-    resume_session_id: str | None = None,
-    resume_checkpoint: dict[str, object] | None = None,
-    idle_output_timeout: int | None = None,
-    ctx: Context = CurrentContext(),
-) -> str:
-    """Dispatch a single food truck L2 session for one recipe.
-
-    Spawns a headless subprocess that executes the given recipe with the
-    provided task and ingredient overrides. Returns a JSON envelope with
-    dispatch_id, dispatched_session_id, l3_payload, and token_usage.
-
-    Args:
-        recipe: Recipe name to dispatch (must be kind=standard).
-        task: Task description for the L2 food truck session.
-        ingredients: Optional ingredient overrides (all values must be strings).
-        dispatch_name: Optional display name for the dispatch record.
-        timeout_sec: Optional L2 session timeout override in seconds.
-        capture: Optional dict mapping capture keys to "${{ result.field }}" templates.
-            Extracted values are persisted in the campaign context for downstream
-            dispatches to reference via "${{ campaign.key }}" in their ingredients.
-        resume_checkpoint: Checkpoint dict from a prior RESUMABLE dispatch envelope.
-            Pass the "resume_checkpoint" field from the prior result to inject completed
-            items context into the resume prompt.
-
-    Never raises.
-    """
-    if (gate := _require_enabled()) is not None:
-        return gate
-    if (fleet_gate := _require_fleet("dispatch_food_truck")) is not None:
-        return fleet_gate
-
-    try:
-        # Feature guard: config authority check independent of MCP visibility state.
-        # Fleet sessions open the gate unconditionally at boot; this catch-all ensures
-        # dispatch_food_truck never executes when features.fleet is disabled in config.
-        from autoskillit.core import FleetErrorCode, fleet_error, is_feature_enabled
-        from autoskillit.server import _get_ctx as _get_ctx_for_feature_check
-
-        _feature_ctx = _get_ctx_for_feature_check()
-        if not is_feature_enabled(
-            "fleet",
-            _feature_ctx.config.features,
-            experimental_enabled=_feature_ctx.config.experimental_enabled,
-        ):
-            return fleet_error(
-                FleetErrorCode.FLEET_FEATURE_DISABLED,
-                "Fleet feature is disabled. Set features.experimental_enabled: true to enable.",
-            )
-
-        campaign_state_path_str = os.environ.get("AUTOSKILLIT_CAMPAIGN_STATE_PATH")
-        continue_on_failure_str = os.environ.get("AUTOSKILLIT_CONTINUE_ON_FAILURE", "false")
-        if campaign_state_path_str and continue_on_failure_str.lower() != "true":
-            from autoskillit.fleet import (  # noqa: PLC0415
-                has_failed_dispatch,
-                reset_failed_dispatch,
-            )
-
-            campaign_sp = Path(campaign_state_path_str)
-            if dispatch_name:
-                reset_failed_dispatch(campaign_sp, dispatch_name)
-            if has_failed_dispatch(campaign_sp):
-                return fleet_error(
-                    FleetErrorCode.FLEET_CAMPAIGN_HALTED,
-                    "Campaign halted: a prior dispatch failed and "
-                    "continue_on_failure is false. "
-                    "No further dispatches permitted.",
-                )
-
-        from autoskillit.core import SessionCheckpoint  # noqa: PLC0415
-        from autoskillit.fleet import execute_dispatch
-        from autoskillit.server import _get_ctx
-        from autoskillit.server._misc import (  # noqa: PLC0415
-            _refresh_quota_cache,
-            check_and_sleep_if_needed,
-            invalidate_cache,
-        )
-
-        parsed_checkpoint = (
-            SessionCheckpoint.from_dict(resume_checkpoint) if resume_checkpoint else None
-        )
-        tool_ctx = _get_ctx()
-        from autoskillit.core import find_caller_session_id
-
-        caller_session_id = find_caller_session_id(project_dir=tool_ctx.project_dir)
-        result = await execute_dispatch(
-            tool_ctx=tool_ctx,
-            recipe=recipe,
-            task=task,
-            ingredients=ingredients,
-            dispatch_name=dispatch_name,
-            timeout_sec=timeout_sec,
-            prompt_builder=_get_food_truck_prompt_builder(),
-            quota_checker=check_and_sleep_if_needed,
-            quota_refresher=_refresh_quota_cache,
-            cache_invalidator=invalidate_cache,
-            capture=capture,
-            resume_session_id=resume_session_id,
-            resume_checkpoint=parsed_checkpoint,
-            idle_output_timeout=idle_output_timeout,
-            caller_session_id=caller_session_id,
-        )
-
-        campaign_state_path_str = os.environ.get("AUTOSKILLIT_CAMPAIGN_STATE_PATH")
-        if campaign_state_path_str and dispatch_name:
-            try:
-                envelope = json.loads(result)
-                campaign_state_path = Path(campaign_state_path_str)
-                if campaign_state_path.exists():
-                    from autoskillit.fleet import (
-                        DispatchRecord,
-                        DispatchStatus,
-                        append_dispatch_record,
-                    )
-
-                    status = DispatchStatus(envelope["dispatch_status"])
-                    append_dispatch_record(
-                        campaign_state_path,
-                        DispatchRecord(
-                            name=dispatch_name,
-                            status=status,
-                            dispatch_id=envelope.get("dispatch_id", ""),
-                            dispatched_session_id=envelope.get("dispatched_session_id", ""),
-                            reason=envelope.get("reason", ""),
-                            token_usage=envelope.get("token_usage") or {},
-                        ),
-                    )
-            except Exception:
-                logger.warning("campaign state update failed", exc_info=True)
-
-        return result
-    except Exception as exc:
-        logger.error("dispatch_food_truck unhandled exception", exc_info=True)
-        from autoskillit.core import FleetErrorCode, fleet_error
-
-        return fleet_error(
-            FleetErrorCode.FLEET_L3_STARTUP_OR_CRASH,
-            f"{type(exc).__name__}: {exc}",
-        )
-
-
-@mcp.tool(
-    tags={"autoskillit", "kitchen-core", "fleet"},
-    annotations={"readOnlyHint": True},
-)
-@track_response_size("record_gate_dispatch")
-async def record_gate_dispatch(
-    dispatch_name: str,
-    approved: bool,
-    ctx: Context = CurrentContext(),
-) -> str:
-    """Record the outcome of a gate dispatch to the campaign state file.
-
-    Gate dispatches are handled by AskUserQuestion (no L3 session). This tool
-    persists the user's approval/rejection so that campaign resume can skip
-    completed gates.
-
-    Args:
-        dispatch_name: Name of the gate dispatch in the campaign manifest.
-        approved: True if the user approved the gate, False if rejected.
-
-    Never raises.
-    """
-    if (gate := _require_enabled()) is not None:
-        return gate
-    if (fleet_gate := _require_fleet("record_gate_dispatch")) is not None:
-        return fleet_gate
-
-    try:
-        from autoskillit.core import FleetErrorCode, fleet_error, is_feature_enabled
-        from autoskillit.fleet import record_gate_outcome
-        from autoskillit.server import _get_ctx as _get_ctx_for_feature_check
-
-        _feature_ctx = _get_ctx_for_feature_check()
-        if not is_feature_enabled(
-            "fleet",
-            _feature_ctx.config.features,
-            experimental_enabled=_feature_ctx.config.experimental_enabled,
-        ):
-            return fleet_error(
-                FleetErrorCode.FLEET_FEATURE_DISABLED,
-                "Fleet feature is disabled. Set features.experimental_enabled: true to enable.",
-            )
-
-        campaign_state_path_str = os.environ.get("AUTOSKILLIT_CAMPAIGN_STATE_PATH")
-        if not campaign_state_path_str:
-            return fleet_error(
-                FleetErrorCode.FLEET_GATE_NO_CAMPAIGN,
-                "No AUTOSKILLIT_CAMPAIGN_STATE_PATH set — not running in campaign mode.",
-            )
-
-        result = record_gate_outcome(Path(campaign_state_path_str), dispatch_name, approved)
-        if not result.success:
-            return fleet_error(FleetErrorCode(result.error_code), result.error_message)
-
-        return json.dumps(
-            {"success": True, "dispatch_name": result.dispatch_name, "status": result.status}
-        )
-    except Exception as exc:
-        logger.error("record_gate_dispatch unhandled exception", exc_info=True)
-        from autoskillit.core import FleetErrorCode, fleet_error
-
-        return fleet_error(
-            FleetErrorCode.FLEET_L3_STARTUP_OR_CRASH,
-            f"{type(exc).__name__}: {exc}",
-        )
